refactor(region): extract CampaignFinanceSyncService — #828 Step 4
Fourth bounded-context split. Extracts the campaign-finance ingestion
pipeline (the CAL-ACCESS streaming download → batch upsert across
contributions / expenditures / independent expenditures / committee
measure filings) out of RegionSyncService.

region-sync.service.ts shrinks from 3,120 → 2,748 LOC (cumulative
3,929 → 2,748, −30% so far). CampaignFinanceSyncService is 457 LOC and
absorbs 5 methods plus 3 local types:

  - syncCampaignFinance (the streaming orchestration)
  - ensureCommitteeStubs (FK reference materialization)
  - sortItems (was sortCampaignFinanceItems — heterogeneous batch routing)
  - upsertBatch (was upsertCampaignFinanceBatch — four-table batch upsert)
  - upsertRecordsByFields (low-level field-projection upsert)
  - PrismaModelDelegate + UpsertConfig + CommitteeRecord (local types)

The orchestrator keeps a thin syncCampaignFinance delegate. Streaming
observability semantics are byte-identical: discover phase opens, the
extract_and_upsert tracker lazily initializes on the first onBatch
callback so phase ordering reads correctly to operators, and the
post-link via PropositionFinanceLinkerService still fires after the
batches complete.

All 2,105 backend tests pass.

Refs #828
